correcciones para implementacion de base de datos

+++ database/migrations/2022_10_18_005345_create_pais_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('paises', function (Blueprint $table) {
             $table->id('id_pais');
-            $table->int('id_region');
+            $table->unsignedBigInteger('id_region');
             $table->string('codigo');
             $table->string('descripcion');
             $table->string('telefono_extension');
