+++ database/migrations/2022_10_18_005635_create_user_metodo_pago_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('usuario_metodo_pago', function (Blueprint $table) {
             $table->id('id_usuario_metodo_pago');
             $table->string('nombre_tarjeta');
-            $table->int('id_metodo_pago');
+            $table->unsignedBigInteger('id_metodo_pago');
             $table->string('codigo_tarjeta');
             $table->string('nombre');
             $table->datetime('fecha_vencimiento');
